Move firebase logic to dedicated folder.
