+++ data/development/2022/04.json
@@ -1,1 +1,1 @@
-[{"area":"dummy-area","address":"http://localhost","addedBy":"iI6agqYmQ2MvjuudmkecHajTifqX","externalId":"dummy","description":"Dummy event description","title":"Dummy event","url":"http://localhost","location":"Online","startTime":"2022-04-16T05:54:36.738Z","endTime":"2022-04-16T06:54:36.738Z","wholeDay":false,"lastModified":"2022-04-15T05:54:36.738Z","category":["dummy-category"],"event":"dummy-event","id":"MuvE8y6twXZVpkLRbgah"}]
+[{"area":"dummy-area","address":"http://localhost","addedBy":"iI6agqYmQ2MvjuudmkecHajTifqX","externalId":"dummy","description":"Dummy event description","title":"Dummy event","url":"http://localhost","location":"Online","startTime":"2022-04-16T05:54:36.738Z","endTime":"2022-04-16T06:54:36.738Z","wholeDay":false,"lastModified":"2022-04-15T05:54:36.738Z","category":["dummy-category"],"event":"dummy-event","id":"MuvE8y6twXZVpkLRbgah"},{"area":"dummy-area","address":"http://localhost","addedBy":"iI6agqYmQ2MvjuudmkecHajTifqX","externalId":"dummy","description":"Dummy event description","title":"Dummy event","url":"http://localhost","location":"Online","startTime":"2022-04-16T05:54:36.738Z","endTime":"2022-04-16T06:54:36.738Z","wholeDay":false,"lastModified":"2022-04-15T05:54:36.738Z","category":["dummy-category"],"event":"dummy-event","id":"MuvE8y6twXZVpkLRbgah"},{"area":"dummy-area","address":"http://localhost","addedBy":"iI6agqYmQ2MvjuudmkecHajTifqX","externalId":"dummy","description":"Dummy event description","title":"Dummy event","url":"http://localhost","location":"Online","startTime":"2022-04-16T15:41:34.976Z","endTime":"2022-04-16T16:41:34.976Z","wholeDay":false,"lastModified":"2022-04-15T15:41:34.976Z","category":["dummy-category"],"event":"dummy-event","id":"l5MjU2Fz1IYntra5eD95"},{"area":"dummy-area","address":"http://localhost","addedBy":"xhKqDWS2NnInQk49l9wT3U5au8Tn","externalId":"dummy","description":"Dummy event description","title":"Dummy event","url":"http://localhost","location":"Online","startTime":"2022-04-16T15:44:45.478Z","endTime":"2022-04-16T16:44:45.478Z","wholeDay":false,"lastModified":"2022-04-15T15:44:45.478Z","category":["dummy-category"],"event":"dummy-event","id":"jSDEYmCbIOKkagHFfOih"}]
