+++ data/development/last-fetch.txt
@@ -1,1 +1,1 @@
-Fri Apr 15 2022 16:13:29 GMT+0000 (Coordinated Universal Time)
+Fri Apr 15 2022 18:44:48 GMT+0300 (Eastern European Summer Time)
