+++ firebase/firestore/test.sh
@@ -3,4 +3,4 @@
 set -e
 
 export FIRESTORE_EMULATOR_HOST="localhost:8080"
-jest -c tools/firebase/jest.config.js --coverage
+jest -c firebase/firestore/jest.config.js --coverage
