+++ jest.config.js
@@ -10,7 +10,7 @@ const customJestConfig = {
   moduleDirectories: ["node_modules", "<rootDir>/"],
   testEnvironment: "jest-environment-jsdom",
   setupFilesAfterEnv: ["<rootDir>/jest.setup.ts"],
-  modulePathIgnorePatterns: ["<rootDir>/tools"],
+  modulePathIgnorePatterns: ["<rootDir>/tools", "<rootDir>/firebase"],
   testPathIgnorePatterns: ["<rootDir>/cypress/"]
 };
