+++ package.json
@@ -10,12 +10,12 @@
     "lint": "next lint && stylelint stylelint **/**.css",
     "start": "next start",
     "test": "jest",
-    "test:db": "firebase emulators:exec --project demo-test tools/firebase/test.sh",
+    "test:db": "firebase emulators:exec --project demo-test firebase/firestore/test.sh",
     "test:export": "./tools/deploy-test/test.sh",
     "cypress": "cypress open",
-    "tools:db:add": "ts-node -O '{\"module\": \"commonjs\"}' ./tools/firebase/add.ts",
-    "tools:db:fetch": "ts-node -O '{\"module\": \"commonjs\"}' ./tools/firebase/fetch.ts",
-    "tools:emulator": "cd tools && firebase emulators:start --project demo-test"
+    "tools:db:add": "ts-node -O '{\"module\": \"commonjs\"}' ./firebase/firestore/add.ts",
+    "tools:db:fetch": "ts-node -O '{\"module\": \"commonjs\"}' ./firebase/firestore/fetch.ts",
+    "tools:emulator": "cd firebase && firebase emulators:start --project demo-test"
   },
   "dependencies": {
     "date-fns": "^2.28.0",
